Add serial terminal
Adds functionality for running openqa-agent against a serial terminal.
The serial terminal function can be used together or instead of the
webserver, but one of either must be present.

Serial terminal works in such a way, that the agent accepts every input
and runs each line as it's own command. The result of the executed is
then serialized as `Reply` json object back via the serial terminal.